+++ pkg/development/devcontext.go
@@ -52,7 +52,7 @@ type DevContext struct {
 
 // NewDevContext creates a new DevContext from the specified request context, parsing and populating
 // the datastore as needed.
-func NewDevContext(ctx context.Context, requestContext *devinterface.RequestContext) (*DevContext, *devinterface.DeveloperErrors, error) {
+func NewDevContext(ctx context.Context, requestContext *devinterface.RequestContext, opts ...CompileOption) (*DevContext, *devinterface.DeveloperErrors, error) {
 	ds, err := memdb.NewMemdbDatastore(0, 0*time.Second, memdb.DisableGC)
 	if err != nil {
 		return nil, nil, err
@@ -60,7 +60,7 @@ func NewDevContext(ctx context.Context, requestContext *devinterface.RequestCont
 	dl := datalayer.NewDataLayer(ds)
 	ctx = datalayer.ContextWithDataLayer(ctx, dl)
 
-	dctx, devErrs, nerr := newDevContextWithDataLayer(ctx, requestContext, dl)
+	dctx, devErrs, nerr := newDevContextWithDataLayer(ctx, requestContext, dl, opts...)
 	if nerr != nil || devErrs != nil {
 		// If any form of error occurred, immediately close the data layer
 		derr := dl.Close()
@@ -74,9 +74,9 @@ func NewDevContext(ctx context.Context, requestContext *devinterface.RequestCont
 	return dctx, nil, nil
 }
 
-func newDevContextWithDataLayer(ctx context.Context, requestContext *devinterface.RequestContext, dl datalayer.DataLayer) (*DevContext, *devinterface.DeveloperErrors, error) {
+func newDevContextWithDataLayer(ctx context.Context, requestContext *devinterface.RequestContext, dl datalayer.DataLayer, opts ...CompileOption) (*DevContext, *devinterface.DeveloperErrors, error) {
 	// Compile the schema and load its caveats and namespaces into the datastore.
-	compiled, devError, err := CompileSchema(requestContext.Schema)
+	compiled, devError, err := CompileSchema(requestContext.Schema, opts...)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -269,6 +269,9 @@ func loadsRels(ctx context.Context, rels []tuple.Relationship, rwt datalayer.Rea
 	return devErrors, err
 }
 
+// loadCompiled validates the compiled schema and then writes it.
+// If validation fails, it returns the full list of validation errors
+// with the correct position.
 func loadCompiled(
 	ctx context.Context,
 	compiled *compiler.CompiledSchema,
@@ -287,66 +290,18 @@ func loadCompiled(
 			continue
 		}
 
-		errWithSource, ok := spiceerrors.AsWithSourceError(cverr)
-		if ok {
-			// NOTE: zeroes are fine here to mean "unknown"
-			lineNumber, err := safecast.Convert[uint32](errWithSource.LineNumber)
-			if err != nil {
-				log.Err(err).Msg("could not cast lineNumber to uint32")
-			}
-			columnPosition, err := safecast.Convert[uint32](errWithSource.ColumnPosition)
-			if err != nil {
-				log.Err(err).Msg("could not cast columnPosition to uint32")
-			}
-			errors = append(errors, &devinterface.DeveloperError{
-				Message: cverr.Error(),
-				Kind:    devinterface.DeveloperError_SCHEMA_ISSUE,
-				Source:  devinterface.DeveloperError_SCHEMA,
-				Context: errWithSource.SourceCodeString,
-				Line:    lineNumber,
-				Column:  columnPosition,
-			})
-		} else {
-			errors = append(errors, &devinterface.DeveloperError{
-				Message: cverr.Error(),
-				Kind:    devinterface.DeveloperError_SCHEMA_ISSUE,
-				Source:  devinterface.DeveloperError_SCHEMA,
-				Context: caveatDef.Name,
-			})
+		if e := getDevError(cverr, compiled, caveatDef); e != nil {
+			errors = append(errors, e)
 		}
 	}
 
 	for _, nsDef := range compiled.ObjectDefinitions {
 		def, terr := schema.NewDefinition(nsDef)
 		if terr != nil {
-			errWithSource, ok := spiceerrors.AsWithSourceError(terr)
-			// NOTE: zeroes are fine here to mean "unknown"
-			lineNumber, err := safecast.Convert[uint32](errWithSource.LineNumber)
-			if err != nil {
-				log.Err(err).Msg("could not cast lineNumber to uint32")
-			}
-			columnPosition, err := safecast.Convert[uint32](errWithSource.ColumnPosition)
-			if err != nil {
-				log.Err(err).Msg("could not cast columnPosition to uint32")
-			}
-			if ok {
-				errors = append(errors, &devinterface.DeveloperError{
-					Message: terr.Error(),
-					Kind:    devinterface.DeveloperError_SCHEMA_ISSUE,
-					Source:  devinterface.DeveloperError_SCHEMA,
-					Context: errWithSource.SourceCodeString,
-					Line:    lineNumber,
-					Column:  columnPosition,
-				})
-				continue
+			if e := getDevError(terr, compiled, nsDef); e != nil {
+				errors = append(errors, e)
 			}
 
-			errors = append(errors, &devinterface.DeveloperError{
-				Message: terr.Error(),
-				Kind:    devinterface.DeveloperError_SCHEMA_ISSUE,
-				Source:  devinterface.DeveloperError_SCHEMA,
-				Context: nsDef.Name,
-			})
 			continue
 		}
 
@@ -356,32 +311,8 @@ func loadCompiled(
 			continue
 		}
 
-		errWithSource, ok := spiceerrors.AsWithSourceError(tverr)
-		if ok {
-			// NOTE: zeroes are fine here to mean "unknown"
-			lineNumber, err := safecast.Convert[uint32](errWithSource.LineNumber)
-			if err != nil {
-				log.Err(err).Msg("could not cast lineNumber to uint32")
-			}
-			columnPosition, err := safecast.Convert[uint32](errWithSource.ColumnPosition)
-			if err != nil {
-				log.Err(err).Msg("could not cast columnPosition to uint32")
-			}
-			errors = append(errors, &devinterface.DeveloperError{
-				Message: tverr.Error(),
-				Kind:    devinterface.DeveloperError_SCHEMA_ISSUE,
-				Source:  devinterface.DeveloperError_SCHEMA,
-				Context: errWithSource.SourceCodeString,
-				Line:    lineNumber,
-				Column:  columnPosition,
-			})
-		} else {
-			errors = append(errors, &devinterface.DeveloperError{
-				Message: tverr.Error(),
-				Kind:    devinterface.DeveloperError_SCHEMA_ISSUE,
-				Source:  devinterface.DeveloperError_SCHEMA,
-				Context: nsDef.Name,
-			})
+		if e := getDevError(tverr, compiled, nsDef); e != nil {
+			errors = append(errors, e)
 		}
 	}
 
@@ -392,6 +323,42 @@ func loadCompiled(
 	}
 
 	return errors, nil
+}
+
+func getDevError(cverr error, compiled *compiler.CompiledSchema, definitionOrCaveat compiler.SchemaDefinition) *devinterface.DeveloperError {
+	if cverr == nil || compiled == nil || definitionOrCaveat == nil {
+		return nil
+	}
+	path := compiled.GetPathToDefinitionOrPartialOrCaveat(definitionOrCaveat.GetName())
+	errWithSource, ok := spiceerrors.AsWithSourceError(cverr)
+	if ok {
+		// NOTE: zeroes are fine here to mean "unknown"
+		// NOTE: positions given by errWithSource are 1-indexed
+		lineNumber, err := safecast.Convert[uint32](errWithSource.LineNumber)
+		if err != nil {
+			log.Err(err).Msg("could not cast lineNumber to uint32")
+		}
+		columnPosition, err := safecast.Convert[uint32](errWithSource.ColumnPosition)
+		if err != nil {
+			log.Err(err).Msg("could not cast columnPosition to uint32")
+		}
+		return &devinterface.DeveloperError{
+			Message: cverr.Error(),
+			Kind:    devinterface.DeveloperError_SCHEMA_ISSUE,
+			Source:  devinterface.DeveloperError_SCHEMA,
+			Context: errWithSource.SourceCodeString,
+			Line:    lineNumber,
+			Column:  columnPosition,
+			Path:    []string{path},
+		}
+	}
+	return &devinterface.DeveloperError{
+		Message: cverr.Error(),
+		Kind:    devinterface.DeveloperError_SCHEMA_ISSUE,
+		Source:  devinterface.DeveloperError_SCHEMA,
+		Context: definitionOrCaveat.GetName(),
+		Path:    []string{path},
+	}
 }
 
 // DistinguishGraphError turns an error from a dispatch call into either a user-facing