+++ pkg/development/schema.go
@@ -2,6 +2,7 @@ package development
 
 import (
 	"errors"
+	"io/fs"
 
 	"github.com/ccoveille/go-safecast/v2"
 
@@ -11,14 +12,49 @@ import (
 	"github.com/authzed/spicedb/pkg/schemadsl/input"
 )
 
+// CompileOption configures schema compilation.
+type CompileOption func(*compileConfig)
+
+type compileConfig struct {
+	fsys         fs.FS
+	rootFileName string
+}
+
+// WithSourceFS enables import resolution using the given filesystem.
+// The filesystem should be rooted at the directory containing the schema file.
+func WithSourceFS(fsys fs.FS) CompileOption {
+	return func(cfg *compileConfig) { cfg.fsys = fsys }
+}
+
+// WithRootFileName saves the root file of the schema
+// so errors can be displayed accurately on UIs.
+func WithRootFileName(name string) CompileOption {
+	return func(cfg *compileConfig) { cfg.rootFileName = name }
+}
+
 // CompileSchema compiles a schema into its caveat and namespace definition(s), returning a developer
 // error if the schema could not be compiled. The non-developer error is returned only if an
 // internal errors occurred.
-func CompileSchema(schema string) (*compiler.CompiledSchema, *devinterface.DeveloperError, error) {
+func CompileSchema(schema string, opts ...CompileOption) (*compiler.CompiledSchema, *devinterface.DeveloperError, error) {
+	cfg := &compileConfig{}
+	for _, o := range opts {
+		o(cfg)
+	}
+
+	var compilerOpts []compiler.Option
+	if cfg.fsys != nil {
+		compilerOpts = append(compilerOpts, compiler.SourceFS(cfg.fsys))
+	}
+
+	sourceFileName := "schema"
+	if cfg.rootFileName != "" {
+		sourceFileName = cfg.rootFileName
+	}
+
 	compiled, err := compiler.Compile(compiler.InputSchema{
-		Source:       input.Source("schema"),
+		Source:       input.Source(sourceFileName),
 		SchemaString: schema,
-	}, compiler.AllowUnprefixedObjectType())
+	}, compiler.AllowUnprefixedObjectType(), compilerOpts...)
 
 	var contextError compiler.WithContextError
 	if errors.As(err, &contextError) {
@@ -43,6 +79,7 @@ func CompileSchema(schema string) (*compiler.CompiledSchema, *devinterface.Devel
 			Line:    uintLine + 1,   // 0-indexed in parser.
 			Column:  uintColumn + 1, // 0-indexed in parser.
 			Context: contextError.ErrorSourceCode,
+			Path:    []string{string(contextError.Source)},
 		}, nil
 	}